==> notebooks/02_Data_Splitting.ipynb <==
# Data Splitting

## Notebook Setup

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split, GroupShuffleSplit
import matplotlib.pyplot as plt
import seaborn as sns

# Import custom functions
import env_functions as ef
import s3_functions as sf

pd.set_option('mode.chained_assignment', None)
pd.options.display.float_format = '{:.2f}'.format
pd.options.display.max_columns = None
pd.options.display.max_rows = None

# Ignore Warnings
import warnings
warnings.simplefilter('ignore', category=FutureWarning)
warnings.simplefilter('ignore', category=UserWarning)

# Determine the environment and get appropriate vars
deepnote, env_vars = ef.load_env_vars()

# Iterate through the vars and set them as global vars
for var_name, var in env_vars.items():
    globals()[var_name] = var

# If not in the DeepNote environment, create a dict for aws creds
#   that were located in the environment file.  This will be passed
#   to all aws s3 functions.
if not deepnote:
    aws_env_vars = {
        'access_key_id': aws_access_key_id,
        'secret_access_key': aws_secret_access_key,
        'bucket_name': s3_bucket_name
    }

## Import Data

if deepnote:
    gcb = pd.read_parquet("/work/data/Global_Coral_Bleaching_DB/gcb_v4.parquet")
else:
    gcb = pd.read_parquet(sf.load_from_s3(file_path="data/Global_Coral_Bleaching_DB/gcb_v4.parquet", **aws_env_vars))

## Drop Imputed Bleaching Data, Non-Hard Coral Data, and Measurements pre-1998

gcb= gcb[(gcb["Substrate_Name"] == "Hard Coral")]
gcb = gcb[gcb["Bleached_Value_Imputed"] == False]
gcb = gcb[gcb["Year"] >= 1998]

## Drop Data Leakage Columns

gcb.drop(columns=['S1','S2','S3','S4','Bleaching_Level','Bleaching_Prevalence_Score',
                  'Severity_ID','Severity_Code','Bleaching_Prevalence_Score_ID',
                  'Percent_Bleached'], inplace=True)

## Perform Split

SEAA_PROVINCE = ['Western Coral Triangle', 'Eastern Coral Triangle', 'Sunda Shelf', 'Tropical Southwestern Pacific',
                 'Northeast Australian Shelf', 'Sahul Shelf', 'South China Sea', 'Andaman', 'Java Transitional']

CARB_PROVINCE = ['Tropical Northwestern Atlantic']

SEAA = gcb[gcb['PROVINCE'].isin(SEAA_PROVINCE)]
CARB = gcb[gcb['PROVINCE'].isin(CARB_PROVINCE)]

# Group split that ensures all observations for a site are in only one split to prevent data leakage

rand_state = 42
tr_size = 0.7
val_size = 0.2
split_size = 1

def custom_shuffle_split(df, tr_size, val_size, rand_state):
    # Shuffle the data
    shuffled_data = df.sample(frac=1, random_state=rand_state)

    # Split shuffled data into train, validation, and test sets
    train_df, remaining_data = train_test_split(shuffled_data, train_size=tr_size, random_state=rand_state)
    holdout_df, val_df = train_test_split(remaining_data, test_size=val_size/(1-tr_size), random_state=rand_state)

    return train_df, val_df, holdout_df

train_df_SEAA, val_df_SEAA, holdout_df_SEAA = custom_shuffle_split(SEAA, tr_size, val_size, rand_state)
train_df_CARB, val_df_CARB, holdout_df_CARB = custom_shuffle_split(CARB, tr_size, val_size, rand_state)
train_df_GLOB, val_df_GLOB, holdout_df_GLOB = custom_shuffle_split(gcb,  tr_size, val_size, rand_state)

## Validation of Distributions

# Validate Depth

fig, axes = plt.subplots(nrows=1, ncols=3, figsize=(12, 4))

alpha = 0.05

# Plot the SEAA data
sns.kdeplot(train_df_SEAA['Depth_m'], color='b', shade=True, alpha=alpha, label='train', ax=axes[0])
sns.kdeplot(val_df_SEAA['Depth_m'], color='g', shade=True, alpha=alpha, label='val', ax=axes[0])
sns.kdeplot(holdout_df_SEAA['Depth_m'], color='r', shade=True, alpha=alpha, label='holdout', ax=axes[0])
axes[0].set_title('SEAA_Depth')
axes[0].legend()

# Plot the CARB data
sns.kdeplot(train_df_CARB['Depth_m'], color='b', shade=True, alpha=alpha, label='train', ax=axes[1])
sns.kdeplot(val_df_CARB['Depth_m'], color='g', shade=True, alpha=alpha, label='val', ax=axes[1])
sns.kdeplot(holdout_df_CARB['Depth_m'], color='r', shade=True, alpha=alpha, label='holdout', ax=axes[1])
axes[1].set_title('CARB_Depth')
axes[1].legend()

# Plot the GLOB data
sns.kdeplot(train_df_GLOB['Depth_m'], color='b', shade=True, alpha=alpha, label='train', ax=axes[2])
sns.kdeplot(val_df_GLOB['Depth_m'], color='g', shade=True, alpha=alpha, label='val', ax=axes[2])
sns.kdeplot(holdout_df_GLOB['Depth_m'], color='r', shade=True, alpha=alpha, label='holdout', ax=axes[2])
axes[2].set_title('GLOB_Depth')
axes[2].legend()

plt.tight_layout()
plt.show()

fig, axes = plt.subplots(nrows=2, ncols=3, figsize=(12, 6))

alpha = 0.05

# Plot the SEAA data -- Long
sns.kdeplot(train_df_SEAA['Longitude_Degrees'], color='b', shade=True, alpha=alpha, label='train', ax=axes[0,0])
sns.kdeplot(val_df_SEAA['Longitude_Degrees'], color='g', shade=True, alpha=alpha, label='val', ax=axes[0,0])
sns.kdeplot(holdout_df_SEAA['Longitude_Degrees'], color='r', shade=True, alpha=alpha, label='holdout', ax=axes[0,0])
axes[0,0].set_title('SEAA_Long')
axes[0,0].legend()

# Plot the CARB data -- Long
sns.kdeplot(train_df_CARB['Longitude_Degrees'], color='b', shade=True, alpha=alpha, label='train', ax=axes[0,1])
sns.kdeplot(val_df_CARB['Longitude_Degrees'], color='g', shade=True, alpha=alpha, label='val', ax=axes[0,1])
sns.kdeplot(holdout_df_CARB['Longitude_Degrees'], color='r', shade=True, alpha=alpha, label='holdout', ax=axes[0,1])
axes[0,1].set_title('CARB_Long')
axes[0,1].legend()

# Plot the GLOB data -- Long
sns.kdeplot(train_df_GLOB['Longitude_Degrees'], color='b', shade=True, alpha=alpha, label='train', ax=axes[0,2])
sns.kdeplot(val_df_GLOB['Longitude_Degrees'], color='g', shade=True, alpha=alpha, label='val', ax=axes[0,2])
sns.kdeplot(holdout_df_GLOB['Longitude_Degrees'], color='r', shade=True, alpha=alpha, label='holdout', ax=axes[0,2])
axes[0,2].set_title('GLOB_Long')
axes[0,2].legend()

# Plot the SEAA data -- Lat
sns.kdeplot(train_df_SEAA['Latitude_Degrees'], color='b', shade=True, alpha=alpha, label='train', ax=axes[1,0])
sns.kdeplot(val_df_SEAA['Latitude_Degrees'], color='g', shade=True, alpha=alpha, label='val', ax=axes[1,0])
sns.kdeplot(holdout_df_SEAA['Latitude_Degrees'], color='r', shade=True, alpha=alpha, label='holdout', ax=axes[1,0])
axes[1,0].set_title('SEAA_Lat')
axes[1,0].legend()

# Plot the CARB data -- Lat
sns.kdeplot(train_df_CARB['Latitude_Degrees'], color='b', shade=True, alpha=alpha, label='train', ax=axes[1,1])
sns.kdeplot(val_df_CARB['Latitude_Degrees'], color='g', shade=True, alpha=alpha, label='val', ax=axes[1,1])
sns.kdeplot(holdout_df_CARB['Latitude_Degrees'], color='r', shade=True, alpha=alpha, label='holdout', ax=axes[1,1])
axes[1,1].set_title('CARB_Lat')
axes[1,1].legend()

# Plot the GLOB data -- Lat
sns.kdeplot(train_df_GLOB['Latitude_Degrees'], color='b', shade=True, alpha=alpha, label='train', ax=axes[1,2])
sns.kdeplot(val_df_GLOB['Latitude_Degrees'], color='g', shade=True, alpha=alpha, label='val', ax=axes[1,2])
sns.kdeplot(holdout_df_GLOB['Latitude_Degrees'], color='r', shade=True, alpha=alpha, label='holdout', ax=axes[1,2])
axes[1,2].set_title('GLOB_Lat')
axes[1,2].legend()

plt.tight_layout()
plt.show()

# Validate Time of Measurement

## split by northern/southern hemisphere first, then month

fig, axes = plt.subplots(nrows=3, ncols=3, figsize=(12, 12))

frames = {'train_SEAA': train_df_SEAA, 'val_SEAA': val_df_SEAA, 'holdout_SEAA': holdout_df_SEAA,
          'train_CARB': train_df_CARB, 'val_CARB': val_df_CARB, 'holdout_CARB': holdout_df_CARB,
          'train_GLOB': train_df_GLOB, 'val_GLOB': val_df_GLOB, 'holdout_GLOB': holdout_df_GLOB}

colors = ['blue', 'green', 'red']
alpha = 0.3

counter = 0
for i in range(3):
    for j in range(3):

        df_name = list(frames.keys())[counter]
        df = frames[df_name]
        month_counts = df.groupby('Date_Month')['Date_Month'].count()
        ax = sns.barplot(x=month_counts.index, y=month_counts.values, ax=axes[i, j], color=colors[i], alpha=alpha)
        axes[i, j].set_title(df_name)

        ax.set_xticks(range(len(month_counts.index)))
        ax.set_xticklabels(month_counts.index.astype(int))
        counter += 1

plt.tight_layout()
plt.show()

## Create X & y Training and Holdout DFs

def X_y_split(train, val, holdout):
    X_train = train.drop(columns='Percent_Bleached_Value')
    y_train = train.Percent_Bleached_Value

    X_val = val.drop(columns='Percent_Bleached_Value')
    y_val = val.Percent_Bleached_Value

    X_holdout = holdout.drop(columns='Percent_Bleached_Value')
    y_holdout = holdout.Percent_Bleached_Value
    return X_train, y_train, X_val, y_val, X_holdout, y_holdout

X_train_SEAA, y_train_SEAA, X_val_SEAA, y_val_SEAA, X_holdout_SEAA, y_holdout_SEAA = X_y_split(train_df_SEAA, val_df_SEAA, holdout_df_SEAA)
X_train_CARB, y_train_CARB, X_val_CARB, y_val_CARB, X_holdout_CARB, y_holdout_CARB = X_y_split(train_df_CARB, val_df_CARB, holdout_df_CARB)
X_train_GLOB, y_train_GLOB, X_val_GLOB, y_val_GLOB, X_holdout_GLOB, y_holdout_GLOB = X_y_split(train_df_GLOB, val_df_GLOB, holdout_df_GLOB)


## Write Out Data

Uncomment the following blocks to write out the data

# Write out SEAA parquet files in DeepNote

# if deepnote:
#     # Local
#     train_df_SEAA.to_parquet('/work/data/Xy_Data/train_df_SEAA.parquet')
#     holdout_df_SEAA.to_parquet('/work/data/Xy_Data/holdout_df_SEAA.parquet')
#     val_df_SEAA.to_parquet('/work/data/Xy_Data/val_df_SEAA.parquet')
#     
#     X_train_SEAA.to_parquet('/work/data/Xy_Data/X_train_SEAA.parquet')
#     y_train_SEAA.to_frame('y_train').to_parquet('/work/data/Xy_Data/y_train_SEAA.parquet')
#     X_val_SEAA.to_parquet('/work/data/Xy_Data/X_val_SEAA.parquet')
#     y_val_SEAA.to_frame('y_val').to_parquet('/work/data/Xy_Data/y_val_SEAA.parquet')
#     X_holdout_SEAA.to_parquet('/work/data/Xy_Data/X_holdout_SEAA.parquet')
#     y_holdout_SEAA.to_frame('y_holdout').to_parquet('/work/data/Xy_Data/y_holdout_SEAA.parquet')
#     
#     # S3
#     train_df_SEAA.to_parquet('/datasets/s3/data/Xy_Data/train_df_SEAA.parquet')
#     holdout_df_SEAA.to_parquet('/datasets/s3/data/Xy_Data/holdout_df_SEAA.parquet')
#     val_df_SEAA.to_parquet('/datasets/s3/data/Xy_Data/val_df_SEAA.parquet')
#     
#     X_train_SEAA.to_parquet('/datasets/s3/data/Xy_Data/X_train_SEAA.parquet')
#     y_train_SEAA.to_frame('y_train').to_parquet('/datasets/s3/data/Xy_Data/y_train_SEAA.parquet')
#     X_val_SEAA.to_parquet('/datasets/s3/data/Xy_Data/X_val_SEAA.parquet')
#     y_val_SEAA.to_frame('y_val').to_parquet('/datasets/s3/data/Xy_Data/y_val_SEAA.parquet')
#     X_holdout_SEAA.to_parquet('/datasets/s3/data/Xy_Data/X_holdout_SEAA.parquet')
#     y_holdout_SEAA.to_frame('y_holdout').to_parquet('/datasets/s3/data/Xy_Data/y_holdout_SEAA.parquet')


# Write out CARB parquet files in DeepNote

# if deepnote:
#     # Local
#     train_df_CARB.to_parquet('/work/data/Xy_Data/train_df_CARB.parquet')
#     holdout_df_CARB.to_parquet('/work/data/Xy_Data/holdout_df_CARB.parquet')
#     val_df_CARB.to_parquet('/work/data/Xy_Data/val_df_CARB.parquet')
#     
#     X_train_CARB.to_parquet('/work/data/Xy_Data/X_train_CARB.parquet')
#     y_train_CARB.to_frame('y_train').to_parquet('/work/data/Xy_Data/y_train_CARB.parquet')
#     X_val_CARB.to_parquet('/work/data/Xy_Data/X_val_CARB.parquet')
#     y_val_CARB.to_frame('y_val').to_parquet('/work/data/Xy_Data/y_val_CARB.parquet')
#     X_holdout_CARB.to_parquet('/work/data/Xy_Data/X_holdout_CARB.parquet')
#     y_holdout_CARB.to_frame('y_holdout').to_parquet('/work/data/Xy_Data/y_holdout_CARB.parquet')
#     
#     # S3
#     train_df_CARB.to_parquet('/datasets/s3/data/Xy_Data/train_df_CARB.parquet')
#     holdout_df_CARB.to_parquet('/datasets/s3/data/Xy_Data/holdout_df_CARB.parquet')
#     val_df_CARB.to_parquet('/datasets/s3/data/Xy_Data/val_df_CARB.parquet')
#     
#     X_train_CARB.to_parquet('/datasets/s3/data/Xy_Data/X_train_CARB.parquet')
#     y_train_CARB.to_frame('y_train').to_parquet('/datasets/s3/data/Xy_Data/y_train_CARB.parquet')
#     X_val_CARB.to_parquet('/datasets/s3/data/Xy_Data/X_val_CARB.parquet')
#     y_val_CARB.to_frame('y_val').to_parquet('/datasets/s3/data/Xy_Data/y_val_CARB.parquet')
#     X_holdout_CARB.to_parquet('/datasets/s3/data/Xy_Data/X_holdout_CARB.parquet')
#     y_holdout_CARB.to_frame('y_holdout').to_parquet('/datasets/s3/data/Xy_Data/y_holdout_CARB.parquet')

# Write out GLOB parquet files in DeepNote

# if deepnote:
#     # Local
#     train_df_GLOB.to_parquet('/work/data/Xy_Data/train_df_GLOB.parquet')
#     holdout_df_GLOB.to_parquet('/work/data/Xy_Data/holdout_df_GLOB.parquet')
#     val_df_GLOB.to_parquet('/work/data/Xy_Data/val_df_GLOB.parquet')
#     
#     X_train_GLOB.to_parquet('/work/data/Xy_Data/X_train_GLOB.parquet')
#     y_train_GLOB.to_frame('y_train').to_parquet('/work/data/Xy_Data/y_train_GLOB.parquet')
#     X_val_GLOB.to_parquet('/work/data/Xy_Data/X_val_GLOB.parquet')
#     y_val_GLOB.to_frame('y_val').to_parquet('/work/data/Xy_Data/y_val_GLOB.parquet')
#     X_holdout_GLOB.to_parquet('/work/data/Xy_Data/X_holdout_GLOB.parquet')
#     y_holdout_GLOB.to_frame('y_holdout').to_parquet('/work/data/Xy_Data/y_holdout_GLOB.parquet')
#     
#     # S3
#     train_df_GLOB.to_parquet('/datasets/s3/data/Xy_Data/train_df_GLOB.parquet')
#     holdout_df_GLOB.to_parquet('/datasets/s3/data/Xy_Data/holdout_df_GLOB.parquet')
#     val_df_GLOB.to_parquet('/datasets/s3/data/Xy_Data/val_df_GLOB.parquet')
#     
#     X_train_GLOB.to_parquet('/datasets/s3/data/Xy_Data/X_train_GLOB.parquet')
#     y_train_GLOB.to_frame('y_train').to_parquet('/datasets/s3/data/Xy_Data/y_train_GLOB.parquet')
#     X_val_GLOB.to_parquet('/datasets/s3/data/Xy_Data/X_val_GLOB.parquet')
#     y_val_GLOB.to_frame('y_val').to_parquet('/datasets/s3/data/Xy_Data/y_val_GLOB.parquet')
#     X_holdout_GLOB.to_parquet('/datasets/s3/data/Xy_Data/X_holdout_GLOB.parquet')
#     y_holdout_GLOB.to_frame('y_holdout').to_parquet('/datasets/s3/data/Xy_Data/y_holdout_GLOB.parquet')

# Write out SEAA parquet files to S3

# if not deepnote:
#     sf.write_to_s3(file_path="data/Xy_Data/train_df_SEAA.parquet", data=train_df_SEAA, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/holdout_df_SEAA.parquet", data=holdout_df_SEAA, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/val_df_SEAA.parquet", data=val_df_SEAA, **aws_env_vars)
#     
#     sf.write_to_s3(file_path="data/Xy_Data/X_train_SEAA.parquet", data=X_train_SEAA, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/y_train_SEAA.parquet", data=y_train_SEAA, **aws_env_vars, series_name='y_train')
#     sf.write_to_s3(file_path="data/Xy_Data/X_val_SEAA.parquet", data=X_val_SEAA, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/y_val_SEAA".parquet, data=y_val_SEAA, **aws_env_vars, series_name='y_val')
#     sf.write_to_s3(file_path="data/Xy_Data/X_holdout_SEAA.parquet", data=X_holdout_SEAA, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/y_holdout_SEAA.parquet", data=y_holdout_SEAA, **aws_env_vars, series_name='y_holdout')


# Write out CARB parquet files to S3

# if not deepnote:
#     sf.write_to_s3(file_path="data/Xy_Data/train_df_CARB.parquet", data=train_df_CARB, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/holdout_df_CARB.parquet", data=holdout_df_CARB, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/val_df_CARB.parquet", data=val_df_CARB, **aws_env_vars)
#     
#     sf.write_to_s3(file_path="data/Xy_Data/X_train_CARB.parquet", data=X_train_CARB, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/y_train_CARB.parquet", data=y_train_CARB, **aws_env_vars, series_name='y_train')
#     sf.write_to_s3(file_path="data/Xy_Data/X_val_CARB.parquet", data=X_val_CARB, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/y_val_CARB".parquet, data=y_val_CARB, **aws_env_vars, series_name='y_val')
#     sf.write_to_s3(file_path="data/Xy_Data/X_holdout_CARB.parquet", data=X_holdout_CARB, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/y_holdout_CARB.parquet", data=y_holdout_CARB, **aws_env_vars, series_name='y_holdout')

# Write out GLOB parquet files to S3

# if not deepnote:
#     sf.write_to_s3(file_path="data/Xy_Data/train_df_GLOB.parquet", data=train_df_GLOB, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/holdout_df_GLOB.parquet", data=holdout_df_GLOB, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/val_df_GLOB.parquet", data=val_df_GLOB, **aws_env_vars)
#     
#     sf.write_to_s3(file_path="data/Xy_Data/X_train_GLOB.parquet", data=X_train_GLOB, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/y_train_GLOB.parquet", data=y_train_GLOB, **aws_env_vars, series_name='y_train')
#     sf.write_to_s3(file_path="data/Xy_Data/X_val_GLOB.parquet", data=X_val_GLOB, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/y_val_GLOB".parquet, data=y_val_GLOB, **aws_env_vars, series_name='y_val')
#     sf.write_to_s3(file_path="data/Xy_Data/X_holdout_GLOB.parquet", data=X_holdout_GLOB, **aws_env_vars)
#     sf.write_to_s3(file_path="data/Xy_Data/y_holdout_GLOB.parquet", data=y_holdout_GLOB, **aws_env_vars, series_name='y_holdout')
